+++ src/managers/macro.py
@@ -1,6 +1,5 @@
 from sc2.ids.unit_typeid import UnitTypeId
 from sc2.unit import Unit
-from units.drone import Drone
 
 
 class MacroManager:
@@ -10,8 +9,11 @@ class MacroManager:
     def __init__(self, bot, structures):
         self.bot = bot
         self.structures = structures
-        self.hq = None
+        # dk why you need to typecast this as a unit
+        self.hq: Unit = self.bot.townhalls.first
         # self.hq = structures.firstexpansion
+        #        self.used_tumors: Set[int] = set()
+        self.inject_interval = 100
 
     async def build_pool(self):  # Build spawning pool
         if (
@@ -22,7 +24,7 @@ class MacroManager:
             if self.can_afford(UnitTypeId.SPAWNINGPOOL):
                 await self.build(
                     UnitTypeId.SPAWNINGPOOL,
-                    near=hq.position.towards(self.game_info.map_center, 5),
+                    near=self.hq.position.towards(self.bot.game_info.map_center, 5),
                 )
 
     async def build_gas(self):
@@ -36,3 +38,33 @@ class MacroManager:
 
     def add_structure(self, unit):
         self.structures = unit
+
+    def _position_blocks_expansion(self, pos):
+        """
+        TODO: figure out why Union and self.expansion_locations_list say they have an error yet no issues arise in the code.  Suspect Pylint is goofed.
+        Note: used pos: Union[Point2, Unit] instead of just pos: Point2 in attempt to fix a y is -1, self.height is 176 error.  Seems to work...
+
+        From Glenn: You don't need to instantiate pos in this function because you're passing a position to this function. Python already knows what it is.
+                    Also, put these docstring inside the function you are referring to.
+                    we should do that for future TODO's as well.
+        """
+
+        blocks_expansion = False
+        for expansion in self.expansion_locations_list:
+            if pos.distance_to(expansion) < 6:
+                blocks_expansion = True
+                break
+        return blocks_expansion
+
+    # Glenn: Idk where this belongs
+    # TODO: Will need to add an array or vector of buildings for "worker_en_route_to_build" to check instead of only HATCHERY
+    # TODO: Check for max number of hatcheries
+
+    async def expand():       
+        # Expands to nearest location when 300 minerals are available up to maximum 5 hatcheries
+        if (
+            self.townhalls.ready.amount + self.already_pending(UnitTypeId.HATCHERY)
+            < 5
+        ):
+            if self.can_afford(UnitTypeId.HATCHERY):
+                await self.expand_now()
